+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/menu/dna_decryptor/DnaDecryptorMenu.kt
@@ -29,7 +29,7 @@ class DnaDecryptorMenu(
 	)
 
 	init {
-		addSlots()
+		addSlots(inventoryY)
 	}
 
 	override fun inputFilter(inputStack: ItemStack): Boolean {